con int anda correctamente Signed-off-by:juanb <juanb@lmde>

+++ application/modules/meetings/models/meeting.php
@@ -62,8 +62,8 @@ class Meeting extends CI_Model {
     }
 
     function get_name($id) {
-        $name = 'XXX ' . (double) $id;
-        $this->db->where(array('id' => (double) $id));
+        $name = 'XXX ' . (int) $id;
+        $this->db->where(array('id' => (int) $id));
         $this->db->select('1693', 'id');
         $rs = $this->db->get($this->container_empresas)->result_array();
         if ($rs)
@@ -72,7 +72,7 @@ class Meeting extends CI_Model {
     }
 
     function get_data($id) {
-        $this->db->where(array('id' => (double) $id));
+        $this->db->where(array('id' => (int) $id));
         $rs = $this->db->get($this->container_empresas)->result_array();
         if ($rs)
             return $rs[0];
@@ -97,20 +97,20 @@ class Meeting extends CI_Model {
     }
 
     function get_empresa($id) {
-        $this->db->where(array('id' => (double) $id));
+        $this->db->where(array('id' => (int) $id));
         $rs = $this->db->get($this->container_empresas)->result_array();
         return $rs;
     }
 
     function accredit($id) {
         $data = array('accredited' => 1);
-        $this->db->where(array('id' => (double) $id));
+        $this->db->where(array('id' => (int) $id));
         $this->db->update($this->container_empresas, $data);
     }
 
     function unaccredit($id) {
         $data = array('accredited' => 0);
-        $this->db->where(array('id' => (double) $id));
+        $this->db->where(array('id' => (int) $id));
         $this->db->update($this->container_empresas, $data);
     }
